Fix builder cache use
We do this by using the docker-container buildx driver rather than the default docker driver and explicitly setting type for the cache-from images.

This also lets us use a newer version of buildkit than is available from the native docker daemon and therefore eliminates the need to annotate Dockerfiles to support caching.